analytics/consumer/requirements.txt
kafka-python-ng==2.2.3
psycopg2-binary==2.9.10
